dsdt: Drop power resources of PCI0.XHC1.RHUB.HS03
Causes the following error since the INT3496 device has been added
for OTG detection:

  ACPI Exception: AE_ERROR, Returned by Handler for [GeneralPurposeIo]
  ACPI Error: Method parse/execution failed \_SB.PCI0.XHC1.RHUB.HS03.WWPR._ON, AE_ERROR

It has questionable usefulness anyway: For some reason, it attempts
to write to the GPIO line that is used as input and interrupt for
the OTG notifications.

Additionally, it seems like GPIO operation regions were not actually
implemented on the stock 3.10 kernel, which means this was never
actually used. Remove it for now to avoid the error shown above.